backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
azure-ai-voicelive[all-websockets]>=1.0.0b5
azure-identity>=1.17.0
